[iOS] feat: ViewModel에 이모지 액션 추가
