+++ iOS/moti/moti/Presentation/Sources/Presentation/GroupDetailAchievement/GroupDetailAchievementViewModel.swift
@@ -11,26 +11,6 @@ import Domain
 import Core
 
 final class GroupDetailAchievementViewModel {
-    enum GroupDetailAchievementViewModelAction {
-        case launch
-        case delete
-        case update(updatedAchievement: Achievement)
-        case blockingAchievement
-        case blockingUser
-    }
-    
-    enum LaunchState {
-        case initial(title: String)
-        case success(achievement: Achievement)
-        case failed(message: String)
-    }
-    
-    enum DeleteState {
-        case loading
-        case success(achievementId: Int)
-        case failed(message: String)
-    }
-
     // MARK: - UseCase
     private let fetchDetailAchievementUseCase: FetchDetailAchievementUseCase
     private let deleteAchievementUseCase: DeleteAchievementUseCase
@@ -44,6 +24,7 @@ final class GroupDetailAchievementViewModel {
     // MARK: - State
     private(set) var launchState = PassthroughSubject<LaunchState, Never>()
     private(set) var deleteState = PassthroughSubject<DeleteState, Never>()
+    private(set) var fetchEmojisState = PassthroughSubject<FetchEmojisState, Never>()
     
     // MARK: - Properties
     private(set) var achievement: Achievement
@@ -87,6 +68,10 @@ final class GroupDetailAchievementViewModel {
             blocking(achievementId: achievement.id)
         case .blockingUser:
             blocking(userCode: achievement.userCode)
+        case .fetchEmojis:
+            fetchEmojis()
+        case .toggleEmoji(let emojiId):
+            toggleEmoji(emojiId)
         }
     }
 }
@@ -141,5 +126,26 @@ extension GroupDetailAchievementViewModel {
     /// 유저를 차단하는 액션
     func blocking(userCode: String) {
         blockingUserUseCase.execute(userCode: userCode)
+    }
+    
+    /// 이모지 리스트를 가져오는 액션
+    func fetchEmojis() {
+        Task {
+            do {
+                let emojis = try await fetchEmojisUseCase.execute(achievementId: achievement.id)
+                fetchEmojisState.send(.success(emojis: emojis))
+            } catch {
+                fetchEmojisState.send(.failed(message: "이모지 정보를 가져오지 못했습니다."))
+            }
+        }
+    }
+    
+    /// 이모지를 토글하는 액션
+    func toggleEmoji(_ emojiId: EmojiType) {
+        Task {
+            // 토글 성공 여부와 상관 없이 Label이 변경됩니다.
+            // 지금은 반환값인 isSuccess를 딱히 쓰지 않아 무시함
+            let _ = try? await toggleEmojiUseCase.execute(achievementId: achievement.id, emojiId: emojiId)
+        }
     }
 }
